requirements.txt
django == 4.2.8
djangorestframework == 3.14.0
python-dateutil == 2.8.2
django-cors-headers == 4.3.1
pytest-django == 4.7.0
pytest == 7.4.4
django-webpack-loader == 3.0.1
